+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-google-client-v1.4.42](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.41...@adobe/spacecat-shared-google-client-v1.4.42) (2025-09-06)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#920](https://github.com/adobe/spacecat-shared/issues/920)) ([1a6b1e1](https://github.com/adobe/spacecat-shared/commit/1a6b1e1ac9531a41c86406ada4bd4ab903307fdc))
+
 # [@adobe/spacecat-shared-google-client-v1.4.41](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.40...@adobe/spacecat-shared-google-client-v1.4.41) (2025-08-09)